+++ docs/cmdline/cluster/src/main.rs
@@ -25,6 +25,7 @@ async fn main() -> Result<()> {
         Some("127.0.0.1".to_owned()),
         Some(8091),
         None,
+        None,
         Some(AgentType::Instance),
     );