+++ docs/cmdline/portal/src/main.rs
@@ -21,6 +21,7 @@ async fn main() -> Result<()> {
         Some("127.0.0.1".to_owned()),
         Some(8090),
         None,
+        None,
         Some(AgentType::Portal),
     );