+++ docs/echo/src/main.rs
@@ -187,6 +187,7 @@ async fn run_client(invitation: &Path) -> Result<(), Error> {
         "127.0.0.1",
         &6502,
         &None,
+        &None,
     )?;
 
     // now give the invitation to connect to the server to the client
@@ -262,7 +263,7 @@ async fn run_server(
     invitation: &Path,
 ) -> Result<(), Error> {
     // create the echo-server service
-    let mut service = ServiceConfig::new("echo-server", url, ip, port, &None)?;
+    let mut service = ServiceConfig::new("echo-server", url, ip, port, &None, &None)?;
 
     let invite = service.add_client("echo-client", range)?;