+++ freeipa/src/main.rs
@@ -45,6 +45,7 @@ async fn main() -> Result<()> {
         Some("127.0.0.1".to_owned()),
         Some(8046),
         None,
+        None,
         Some(AgentType::Account),
     );