+++ provider/src/main.rs
@@ -41,6 +41,7 @@ async fn main() -> Result<()> {
         Some("127.0.0.1".to_owned()),
         Some(8041),
         None,
+        None,
         Some(AgentType::Provider),
     );