+++ templemeads/src/agent_core.rs
@@ -74,6 +74,7 @@ pub struct Defaults {
 }
 
 impl Defaults {
+    #[allow(clippy::too_many_arguments)]
     pub fn parse(
         name: Option<String>,
         config_file: Option<PathBuf>,
@@ -81,10 +82,19 @@ impl Defaults {
         ip: Option<String>,
         port: Option<u16>,
         healthcheck_port: Option<u16>,
+        proxy_header: Option<String>,
         agent: Option<AgentType>,
     ) -> Self {
         Self {
-            service: ServiceDefaults::parse(name, config_file, url, ip, port, healthcheck_port),
+            service: ServiceDefaults::parse(
+                name,
+                config_file,
+                url,
+                ip,
+                port,
+                healthcheck_port,
+                proxy_header,
+            ),
             agent: agent.unwrap_or(AgentType::Portal),
             extras: HashMap::new(),
         }
@@ -132,6 +142,7 @@ pub async fn process_args(defaults: &Defaults) -> Result<Option<Config>, Error> 
             ip,
             port,
             healthcheck_port,
+            proxy_header,
             force,
         }) => {
             let local_healthcheck_port;
@@ -153,6 +164,7 @@ pub async fn process_args(defaults: &Defaults) -> Result<Option<Config>, Error> 
                             .to_string(),
                         &port.unwrap_or_else(|| defaults.service.port()),
                         &local_healthcheck_port,
+                        proxy_header,
                     )?
                 },
                 agent: defaults.agent.clone(),
@@ -392,6 +404,13 @@ enum Commands {
             help = "Optional port on which to listen for health checks (e.g. 8080)"
         )]
         healthcheck_port: Option<u16>,
+
+        #[arg(
+            long,
+            short = 'x',
+            help = "Proxy header to use for the client IP address - look here for the client IP address"
+        )]
+        proxy_header: Option<String>,
 
         #[arg(long, short = 'f', help = "Force reinitialisation")]
         force: bool,